requirements.txt
annotated-types==0.6.0
anyio==4.6.2
APScheduler==3.11.0
blinker==1.8.2
brotlicffi==1.0.9.2
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
distro==1.9.0
Flask==3.0.3
Flask-APScheduler==1.13.1
gevent==25.4.1
greenlet==3.2.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.79
langchain-ollama==0.3.10
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.32
MarkupSafe==3.0.2
Naked==0.1.32
numpy==2.3.4
ollama==0.6.0
openai==2.6.1
orjson==3.10.14
packaging==25.0
pip==25.2
psycopg2-binary==2.9.11
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.2
pydantic_core==2.41.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.2
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
setuptools==80.9.0
shellescape==3.8.1
six==1.17.0
sniffio==1.3.0
SQLAlchemy==2.0.43
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
zope.event==6.0
zope.interface==8.0.1
zstandard==0.24.0